Moved zim::Reader to a header file of its own
