+++ src/file_reader.h
@@ -20,53 +20,11 @@
 #ifndef ZIM_FILE_READER_H_
 #define ZIM_FILE_READER_H_
 
-#include <memory>
-
-#include "zim_types.h"
-#include "endian_tools.h"
-#include "debug.h"
+#include "reader.h"
 
 namespace zim {
 
-class Buffer;
 class FileCompound;
-
-class Reader {
-  public:
-    Reader() {};
-    virtual zsize_t size() const = 0;
-    virtual ~Reader() {};
-
-    virtual void read(char* dest, offset_t offset, zsize_t size) const = 0;
-    template<typename T>
-    T read_uint(offset_t offset) const {
-      ASSERT(offset.v, <, size().v);
-      ASSERT(offset.v+sizeof(T), <=, size().v);
-      char tmp_buf[sizeof(T)];
-      read(tmp_buf, offset, zsize_t(sizeof(T)));
-      return fromLittleEndian<T>(tmp_buf);
-    }
-    virtual char read(offset_t offset) const = 0;
-
-    virtual std::shared_ptr<const Buffer> get_buffer(offset_t offset, zsize_t size) const = 0;
-    std::shared_ptr<const Buffer> get_buffer(offset_t offset) const {
-      return get_buffer(offset, zsize_t(size().v-offset.v));
-    }
-    virtual std::unique_ptr<const Reader> sub_reader(offset_t offset, zsize_t size) const = 0;
-    std::unique_ptr<const Reader> sub_reader(offset_t offset) const {
-      return sub_reader(offset, zsize_t(size().v-offset.v));
-    }
-    virtual offset_t offset() const = 0;
-
-    std::unique_ptr<const Reader> sub_clusterReader(offset_t offset,
-                                                    CompressionType* comp,
-                                                    bool* extented) const;
-
-    bool can_read(offset_t offset, zsize_t size);
-
-  private:
-    std::shared_ptr<const Buffer> get_clusterBuffer(offset_t offset, CompressionType comp) const;
-};
 
 class FileReader : public Reader {
   public:
